+++ test/unit/test-helper.ts
@@ -18,7 +18,8 @@
 
 import {
   CredentialStatus,
-  IProofParams,
+  IChallengeRequestParams,
+  ISecp256k1ProofParams,
   IVerifiableCredentialParams,
   IVerifiablePresentationParams,
   VerifiableCredential
@@ -27,14 +28,14 @@ import {
 /**
  * Proof containing a type, created date and verification method with static, fictive data
  */
-export const testProofParams: IProofParams = {
+export const testProofParams: ISecp256k1ProofParams = {
   type: 'SignatureAlgorithmName',
   created: new Date('01-01-2019'),
   verificationMethod: 'verification method',
   nonce: 'f8bdf225-a6b3-49df-8d6a-201943a0b36d'
 }
 
-export const challengeRequestTestData = {
+export const challengeRequestTestData: IChallengeRequestParams = {
   toAttest: [
     { predicate: 'https://schema.org/FinancialProduct' }
   ],
@@ -63,7 +64,7 @@ const testCredStatus = new CredentialStatus({
   type: 'vcStatusRegistry2019'
 })
 
-export const testCredentialParams = {
+export const testCredentialParams: IVerifiableCredentialParams = {
   id: 'did:protocol:address',
   type: ['VerifiableCredential'],
   issuer: 'did:protocol:issueraddress',
@@ -75,13 +76,13 @@ export const testCredentialParams = {
   proof: {
     type: 'SignatureAlgorithmName',
     created: new Date('01-01-2019'),
-    verificationMethod: 'verification method',
+    verificationMethod: 'verification method X',
     nonce: '547d06de-7f1b-4040-8ad0-cbee414a4a7f'
   },
   credentialStatus: testCredStatus,
   optionalField: 'optionalContent',
   '@context': ['https://www.w3.org/2018/credentials/v1', 'https://schema.org/givenName']
-} as IVerifiableCredentialParams
+}
 
 export const testCredential = new VerifiableCredential(testCredentialParams)